requirements.txt
Flask==3.0.0
openai==1.54.0
httpx==0.27.0
markdown==3.5.1
gunicorn==21.2.0

python-dotenv==1.0.0
